#04 - Test lançar erro de validação quando não houver dados suficientes para criação do livros

+++ src/test/java/com/wennersanner/libraryapi/service/BookServiceTest.java
@@ -38,7 +38,7 @@ public class BookServiceTest {
 
         Mockito.when( repository.save(book)).thenReturn(
                         Book.builder()
-                                .id(1)
+                                .id(Long.valueOf((long) 1))
                                 .author("Fulano")
                                 .isbn("123")
                                 .title("As aventuras").build());
